feat: added exception handling
Made swissPairings function raise a RuntimeError exception in case trials
exceed a limit. The exception is caught in the 'populate.py' module which
causes the app to try to randomly match pairings a few more times before the
app quits.
This is to avoid the situation where swissPairings gets stuck in an infinite
loop.